Lab-1

+++ ЛБ2_05.03.22/Form1.cs
@@ -19,46 +19,46 @@ namespace ЛБ2_05._03._22
 
         private void button1_Click(object sender, EventArgs e) // сложение
         {
-            double a, b;
-            a = Double.Parse(textBox1.Text);
-            b = Double.Parse(textBox2.Text);
+            int a, b;
+            a = Int32.Parse(textBox1.Text);
+            b = Int32.Parse(textBox2.Text);
 
-            double c = a + b;
+            int c = a + b;
             textBox3.Text = c.ToString();
             label5.Text = "Сложение";
         }
 
         private void button2_Click(object sender, EventArgs e) // вычитание
         {
-            double a, b;
-            a = Double.Parse(textBox1.Text);
-            b = Double.Parse(textBox2.Text);
+            int a, b;
+            a = Int32.Parse(textBox1.Text);
+            b = Int32.Parse(textBox2.Text);
 
-            double c = a - b;
+            int c = a - b;
             textBox3.Text = c.ToString();
             label5.Text = "Вычитание";
         }
 
         private void button3_Click(object sender, EventArgs e) // умножение 
         {
-            double a, b;
-            a = Double.Parse(textBox1.Text);
-            b = Double.Parse(textBox2.Text);
+            int a, b;
+            a = Int32.Parse(textBox1.Text);
+            b = Int32.Parse(textBox2.Text);
 
-            double c = a * b;
+            int c = a * b;
             textBox3.Text = c.ToString();
             label5.Text = "Умножение";
         }
 
         private void button4_Click(object sender, EventArgs e) // деление
         {
-            double a, b;
-            a = Double.Parse(textBox1.Text);
-            b = Double.Parse(textBox2.Text);
+            int a, b;
+            a = Int32.Parse(textBox1.Text);
+            b = Int32.Parse(textBox2.Text);
 
             if(b != 0)
             {
-                double c = a / b;
+                int c = a / b;
                 textBox3.Text = c.ToString();
                 label5.Text = "Деление";
             }
